requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
chromadb>=0.4.18
ollama>=0.1.7
PyYAML>=6.0.1
pypdf>=3.17.0
sentence-transformers>=2.2.2
tiktoken>=0.5.1
watchdog>=3.0.0
psutil>=5.9.0
requests>=2.28.0